Restrict “current” directory usage
This removes `setCurrentDirectory` and moves `withCurrentDirectory` to a
`Compat` module (because it can be very helpful when calling external code that
has expectations around using the current directory).

It also improves the relevant docs a bit.
