+++ system/pathway-system.cabal
@@ -134,6 +134,9 @@ library
     transformers ^>= {0.5.6, 0.6.1},
   exposed-modules:
     Filesystem.Path
+    Filesystem.Path.Compat
+  other-modules:
+    Filesystem.Path.Internal
 
 test-suite doctests
   import: defaults
